requirements.txt
# requirements.txt
torch
clip-by-openai
faiss-gpu
gradio
numpy
tqdm
Pillow
